fix(intake): phone_lock relocated to light backend package + remaining test-double gaps
The pre-push full suite caught two gaps the targeted battery missed:
- backend.db/__init__ and backend.utils/__init__ both drag
  backend.app.core.config at import (OPENAI warning) — the reprocess script's
  --help must stay settings-free, so phone_lock moves to backend/phone_lock.py
  (the bare `backend` package imports in 149us with zero side effects).
  All 7 consumer imports updated; script --help verified warning-free.
- ClientIdentityResolver (assignment.py) FakeConn in
  test_client_identity_resolver.py taught .transaction() for the F12 lock TX.

Battery: 695 passed, 1 skipped.
