+++ apps/backend-rag/backend/services/crm/cache_query.py
@@ -16,7 +16,7 @@ from typing import Any, TypeVar
 
 import asyncpg
 
-from backend.db.phone_lock import lock_phone_cores
+from backend.phone_lock import lock_phone_cores
 from backend.services.common.background import spawn
 
 logger = logging.getLogger(__name__)
